+++ src/applications/disability-benefits/disability-rating-calculator/components/CalculatedDisabilityRating.jsx
