+++ src/applications/disability-benefits/disability-rating-calculator/components/RatingRow.jsx
