+++ src/applications/disability-benefits/disability-rating-calculator/sass/disability-calculator.scss
