+++ src/applications/disability-benefits/disability-rating-calculator/tests/utils/inputs.json
