+++ src/applications/disability-benefits/disability-rating-calculator/utils/helpers.js
